+++ src/System.Collections.Concurrent/ref/System.Collections.Concurrent.cs
@@ -72,6 +72,7 @@ namespace System.Collections.Concurrent
         public T[] ToArray() { throw null; }
         public bool TryPeek(out T result) { throw null; }
         public bool TryTake(out T result) { throw null; }
+        public void Clear() { throw null; }
     }
     public partial class ConcurrentDictionary<TKey, TValue> : System.Collections.Generic.ICollection<System.Collections.Generic.KeyValuePair<TKey, TValue>>, System.Collections.Generic.IDictionary<TKey, TValue>, System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<TKey, TValue>>, System.Collections.Generic.IReadOnlyCollection<System.Collections.Generic.KeyValuePair<TKey, TValue>>, System.Collections.Generic.IReadOnlyDictionary<TKey, TValue>, System.Collections.ICollection, System.Collections.IDictionary, System.Collections.IEnumerable
     {
