+++ src/System.Collections.Concurrent/tests/System.Collections.Concurrent.Tests.csproj
@@ -89,9 +89,6 @@
       <Link>Common\System\Diagnostics\DebuggerAttributes.cs</Link>
     </Compile>
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp1.1'">
-    <Compile Include="ConcurrentBagTests.netcoreapp1.1.cs" />
-  </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)'!=''">
     <Compile Include="$(CommonPath)\System\SerializableAttribute.cs">
       <Link>Common\System\SerializableAttribute.cs</Link>
@@ -104,6 +101,7 @@
     <Compile Include="$(CommonTestPath)\System\Collections\IEnumerable.Generic.Serialization.Tests.cs">
       <Link>Common\System\Collections\IEnumerable.Generic.Serialization.Tests.cs</Link>
     </Compile>
+    <Compile Include="ConcurrentBagTests.netstandard1.7.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\pkg\System.Collections.Concurrent.pkgproj">
